Bug fix to provide a better SD fit parameter for the pileup cut.
There was a problem with one of the data runs where the pileup cut was
way too narrow. The initial guess for the standard deviation of the
gaussian was too small, so it did not fit properly. Because the
previous iteration is used in the fit, this created problems further
on. This commit also fixes a small bug where the pilup cut reconnects
at y_top instead of y_top_new. This shouldn't actually make a
difference because the smoothing does not affect the first few points,
but this code is more correct.